requirements.txt
fastapi
uvicorn
python-dotenv
openai-agents
tiktoken
# Django integration dependencies
django
psycopg2-binary
redis
django-cors-headers
anthropic